projeto_10810/requirements.txt
pandas
numpy
matplotlib
pillow
openpyxl